Upgrade to latest XDelta version, 3.1.0
Apparently 3.1.0 is considered a beta version, but it's been the newest version for the last 4 years so it's good enough for me. Per the notes for that version, it just has some slight performance loss compared to the last stable release, but the version we were on before this is so old that 3.1.0 is still faster and better in every way.